Update to the SDK track 2

+++ ConnectionsManagement/Prepare.cs
@@ -1,18 +1,17 @@
-namespace ConnectionsManagement
-{
-    using System.Threading.Tasks;
-    using Microsoft.Azure.ServiceBus.Management;
+using System.Threading.Tasks;
+using Azure.Messaging.ServiceBus.Administration;
 
-    public static class Prepare
+namespace AtomicSends;
+
+public static class Prepare
+{
+    public static async Task Infrastructure(string connectionString, string destination)
     {
-        public static async Task Infrastructure(string connectionString, string destination)
+        var client = new ServiceBusAdministrationClient(connectionString);
+        if (await client.QueueExistsAsync(destination))
         {
-            var client = new ManagementClient(connectionString);
-            {
-                await client.DeleteQueueAsync(destination);
-            }
-            await client.CreateQueueAsync(destination);
-            await client.CloseAsync();
+            await client.DeleteQueueAsync(destination);
         }
+        await client.CreateQueueAsync(destination);
     }
 }
